+++ pyroomacoustics/__init__.py
@@ -67,7 +67,7 @@ Available submodules
     Abstraction for a sound source.
 
 :py:obj:`pyroomacoustics.stft`
-    **Deprecated** Replaced by the methods in :py:obj:`pyroomacoustics.realtime`
+    **Deprecated** Replaced by the methods in :py:obj:`pyroomacoustics.transform`
 
 :py:obj:`pyroomacoustics.sync`
     A few routines to help synchronize signals.
@@ -96,7 +96,7 @@ Available subpackages
 :py:obj:`pyroomacoustics.doa`
     Direction of arrival finding algorithms
 
-:py:obj:`pyroomacoustics.realtime`
+:py:obj:`pyroomacoustics.transform`
     Block frequency domain processing tools
 
 
@@ -126,7 +126,14 @@ from .recognition import *
 
 from . import doa
 from . import adaptive
-from . import realtime
+from . import transform
+from . import transform as realtime  # to be deprecated
 from . import experimental
 from . import datasets
 from . import bss
+
+import warnings
+warnings.warn(
+    "Do not import the 'realtime' module as it will be deprecated soon. It will be replaced by the 'transform' module."
+    , DeprecationWarning
+)