+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 Added
 ~~~~~
 
+- Renamed ``realtime`` to ``transform`` and added deprecation warning.
 - Added a cython function to efficiently compute the fractional delays in the room
   impulse response from time delays and attenuations
 - `notebooks` folder.